Start rbd-target-api when rbd-target-gw starts
rbd-target-api is bound to rbd-target-gw to prevent issues where
rbd-target-gw goes down but rbd-target-api stays up and continues to
make changes to the system and they then race making changes when
rbd-target-gw is brought back up. The BindTo setting makes sure that
rbd-target-api goes down with rbd-target-gw. The problem is that if you
bring up rbd-target-gw, rbd-target-api is not automatically brought back
up. This adds the Wants setting so it is always brought up with
rbd-target-gw.